First Git Commit
Before, I was only committing to a local repo. My bad. The site is functional (muscles are clicked on, light up, are recorded to that user, and can then be visualized when 'gaze'd at), and has a crude new-user page. Next important steps: style the new-user page, add an add-new-user button, and eventually get the mouseover events to work properly.
Database is still sqlite.

Signed-off-by: Alexander Loewi <amloewi@gmail.com>